refactor(proxy): split streaming retry mixin
